+++ pygal/graph/stackedline.py
@@ -39,6 +39,22 @@ class StackedLine(Line):
         self._previous_line = None
         super(StackedLine, self).__init__(*args, **kwargs)
 
+    def _value_format(self, value, serie, index):
+        """
+        Display value and cumulation
+        """
+        sum_ = serie.points[index][1]
+        if serie in self.series and (
+                self.stack_from_top and
+                self.series.index(serie) == self._order - 1 or
+                not self.stack_from_top and
+                self.series.index(serie) == 0):
+            return super(StackedLine, self)._value_format(value)
+        return '%s (+%s)' % (
+            self._y_format(sum_),
+            self._y_format(value)
+        )
+
     def _fill(self, values):
         """Add extra values to fill the line"""
         if not self._previous_line: